Adding winner method
Have added a method to determine who is the winner. Trying to figure out how best to test this option. With me having to create a loop to get a winner, trying to see if there is a way to mock player 1 or 2 and the array to test it out