Updating Board Class
Adding a way for the board to collect and store the position of X and O on the board before printing. Went through a few different options, and original was a bit longer, but thanks to some help I was able to get to the one I have now. Have tested some of it but not been able to test the elsif statement in user choice